feat(shared): Complete response body parser
features:
- Multi-format support (byte[], String, Object)
- Error response parsing
- Collection type handling
- Content-Type aware charset conversion

changes
- rename to ResponseBodyParser
